Convert Game to TypeScript
